pyproject.toml
[project]
name = "medical-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.17",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.1.2",
    "loguru>=0.7.3",
    "pinecone>=8.1.0",
    "pydantic>=2.12.5",
    "pypdf>=6.7.5",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "tqdm>=4.67.3",
    "uvicorn>=0.41.0",
]
